switched from glfw to SDL2

+++ include/core/Clock.h
@@ -15,8 +15,8 @@ public:
 	Time getTime() const;
 
 private:
-	double	m_startTime;
-	double	m_stopTime;
+	uint32	m_startTime;
+	uint32  m_stopTime;
 	bool	m_running;
 	bool	m_reset;
 };
